refactor(wallet): Replace lookup for AOCL index with StrongKey
Replace the AOCL leaf index lookup mapping in the wallet database with a
`StrongUtxoKey` lookup mapping. This means that you now need the
addition record and the AOCL leaf index to probe the wallet database for
element inclusion of a monitored UTXO. This is done since this is the way
we do duplication checks in the wallet database -- this is how we probe
if the wallet database already contains the specified UTXO/monitored
UTXO.

This simplifies some logic in the wallet state updating function
`WalletState::update_wallet_state_with_new_block` since we now don't
have to precalculate the potential duplicates -- since the database
logic can handle duplicate-checking internally. Database migration to v2
has been adapted to this schema.